Improve error handling on JSON pase failures on success messages
Previously the library would report a network failure if the response
to an HTTP request couldn't be parsed either as a success value or an
error value. This commit improves the situation, by only reporting a
network failure if the HTTP response code is not successful and
reporting the JSON parse error (of the success value) otherwise.